+++ rest/src/main/java/com/itwill/rest/repository/SongFirstConRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.itwill.rest.domain.SongFirstCon;
 
-public interface SongFirstConRepository extends JpaRepository<SongFirstCon, Long>{
+//public interface SongFirstConRepository extends JpaRepository<SongFirstCon, Long>{
+public interface SongFirstConRepository extends JpaRepository<SongFirstCon, Integer>{
 	
 }
